# Conflicts:
#	AGXUnrealDev/Plugins/AGXUnreal/Source/AGXUnreal/Private/AGX_RigidBodyComponent.cpp
#	AGXUnrealDev/Plugins/AGXUnreal/Source/AGXUnrealEditor/Public/Utilities/AGX_EditorUtilities.h
#	AGXUnrealDev/Plugins/AGXUnreal/Source/AGXUnrealTests/Private/AGX_ImporterToBlueprintTest.cpp

+++ Source/AGXUnrealBarrier/Public/Shapes/HeightFieldShapeBarrier.h
@@ -17,7 +17,7 @@ public:
 	FHeightFieldShapeBarrier& operator=(FHeightFieldShapeBarrier&& Other) = default;
 
 	void AllocateNative(
-		int32 NumVerticesX, int32 NumVerticesY, float SizeX, float SizeY,
+		int32 NumVerticesX, int32 NumVerticesY, double SizeX, double SizeY,
 		const TArray<float>& Heights);
 
 private:
@@ -25,7 +25,7 @@ private:
 	virtual void ReleaseNativeShape() override;
 
 	void AllocateNativeHeightField(
-		int32 NumVerticesX, int32 NumVerticesY, float SizeX, float SizeY,
+		int32 NumVerticesX, int32 NumVerticesY, double SizeX, double SizeY,
 		const TArray<float>& Heights);
 
 private:
